restructure runner code
